Relocate SetAutonaming
